+++ src/components/main-page/MainPage.test.tsx
@@ -1,7 +1,7 @@
 import { describe, expect } from 'vitest';
-import MainPage from '../components/MainPage.tsx';
+import MainPage from './MainPage.tsx';
 import { MemoryRouter } from 'react-router-dom';
-import { renderWithProviders } from './test-utils.tsx';
+import { renderWithProviders } from '../../test/test-utils.tsx';
 
 describe('Main page.tsx', () => {
   it('should define component', function () {
