+++ src/components/person-details/Details.test.tsx
@@ -1,14 +1,14 @@
 import { afterEach, beforeEach, describe, expect } from 'vitest';
 import { screen, waitFor } from '@testing-library/react';
-import Details from '../components/Details.tsx';
+import Details from './Details.tsx';
 import { BrowserRouter, MemoryRouter, Route, Routes } from 'react-router-dom';
 import { userEvent } from '@testing-library/user-event';
-import { renderWithProviders } from './test-utils.tsx';
-import { ROUTER_PATHS } from '../router/router.tsx';
-import { PERSON_MOCK } from '../models/mock.const.tsx';
-import App from '../App.tsx';
-import { server } from './server.tsx';
-import { API_URL } from '../services/apiSlice.tsx';
+import { renderWithProviders } from '../../test/test-utils.tsx';
+import { ROUTER_PATHS } from '../../router/router.tsx';
+import { PERSON_MOCK } from '../../models/mock.const.tsx';
+import App from '../../App.tsx';
+import { server } from '../../test/server.tsx';
+import { API_URL } from '../../store/api/apiSlice.tsx';
 
 describe('Details.tsx interactions', () => {
   const user = userEvent.setup();
@@ -30,7 +30,7 @@ describe('Details.tsx interactions', () => {
   });
 
   //Validate that clicking on a card opens a detailed card component;
-  it('should show details component when click on person card', async () => {
+  it('should show person-details component when click on person card', async () => {
     await waitFor(() => {
       user.click(screen.getAllByTestId('person-container')[0]);
     });
@@ -39,7 +39,7 @@ describe('Details.tsx interactions', () => {
   });
 
   //Ensure that clicking the close button hides the component
-  it('should hide details component on btn close click', async () => {
+  it('should hide person-details component on btn close click', async () => {
     await waitFor(async () => {
       await user.click(screen.getAllByTestId('person-container')[0]);
       const closeButton = screen.getByTestId('detail-close-btn');
